Set correct mock api path
