+++ cypress/e2e/comparator.cy.ts
@@ -1,4 +1,4 @@
-import { getApiBaseUrl } from '~/github-client'
+import { getMockApiPath } from '~/mocks/utils'
 
 const DEFAULT_COMMAND_TIMEOUT = Cypress.config('defaultCommandTimeout')
 const LONGER_COMMAND_TIMEOUT = DEFAULT_COMMAND_TIMEOUT * 5
@@ -239,7 +239,7 @@ it('should show changelog results when preloading from URL with more than 10 rel
 
 			worker.use(
 				rest.get(
-					`${getApiBaseUrl()}/repos/renovatebot/renovate/releases`,
+					`${getMockApiPath()}/repos/renovatebot/renovate/releases`,
 					(req, res) => {
 						const pageIndex = Number(req.url.searchParams.get('page') || 1)
