+++ src/mocks/handlers/github-repos-releases.ts
@@ -1,11 +1,11 @@
-import { type RequestHandler, type DefaultBodyType } from 'msw'
+import { type DefaultBodyType, type RequestHandler } from 'msw'
 import { rest } from 'msw'
 
 import {
 	domTestingLibraryReleases,
 	renovateReleases,
 } from '@/fixtures/github/releases'
-import { getApiBaseUrl } from '@/github-client'
+import { getMockApiPath } from '@/mocks/utils'
 import type { Release } from '@/models'
 import { paginateList } from '@/utils'
 
@@ -26,7 +26,7 @@ const REPO_FIXTURES_MAPPING: Record<string, Array<Release> | undefined> = {
 
 const githubReposReleasesHandlers: Array<RequestHandler> = [
 	rest.get<DefaultBodyType, RepoReleasesParams>(
-		`${getApiBaseUrl()}/repos/:repoOwner/:repoName/releases`,
+		`${getMockApiPath()}/repos/:repoOwner/:repoName/releases`,
 		(req, res, context) => {
 			const { repoOwner, repoName } = req.params
 			const releasesFixture = REPO_FIXTURES_MAPPING[repoName]
@@ -60,7 +60,7 @@ const githubReposReleasesHandlers: Array<RequestHandler> = [
 				responseTransformers.push(
 					context.set(
 						'link',
-						`<${getApiBaseUrl()}/repos/${repoString}/releases?per_page=${perPage}&page=${nextPage}>; rel="next"`,
+						`<${getMockApiPath()}/repos/${repoString}/releases?per_page=${perPage}&page=${nextPage}>; rel="next"`,
 					),
 				)
 			}
