+++ src/mocks/index.ts
@@ -1,22 +1,6 @@
-import { rest, type StartOptions } from 'msw'
+import { rest } from 'msw'
 
-const IGNORE_HOSTS = ['localhost', 'octoclairvoyant', 'fonts']
-
-type OnUnhandledRequestCallback = StartOptions['onUnhandledRequest']
-
-const unhandledRequestCallback: OnUnhandledRequestCallback = (req, print) => {
-	if (req.url.host === 'api.github.com') {
-		throw new Error(
-			`Unhandled request to GitHub API: ${req.method.toUpperCase()} ${req.url.toString()}`,
-		)
-	}
-
-	if (IGNORE_HOSTS.some((ignoreHost) => req.url.host.includes(ignoreHost))) {
-		return undefined
-	}
-
-	print.warning()
-}
+import { unhandledRequestCallback } from './utils'
 
 async function initMocks(): Promise<ServiceWorkerRegistration | undefined> {
 	const isServerEnv = typeof window === 'undefined'
